requirements-base.txt
# ============================================================================
# Base Requirements - Production & Development共通
# ============================================================================

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
cloudscraper==1.2.71

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0
joblib==1.3.2

# Database
psycopg2-binary==2.9.9

# Task Scheduling
APScheduler==3.10.4
schedule==1.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
